fix: redundant types
