+++ packages/core/index.ts
@@ -15,3 +15,4 @@ export * from './agent-message';
 export * from './types/api';
 export * from './types/utils';
 export * from './types/analytics';
+export * from './types/agent-snapshots';
