+++ packages/core/types/agent-snapshots.ts
@@ -1,0 +1,28 @@
+export type AgentSnapshotMetadata = {
+  traceId: string;
+  iterations: AgentSnapshotIterationMetadata[];
+};
+
+export type AgentSnapshotIterationMetadata = {
+  iterationNumber: number;
+  folderName: string;
+  timestamp: string;
+  jsonFileCount: number;
+};
+
+export type AgentSnapshotFolder = {
+  folderName: string;
+  fullPath: string;
+  traceId: string;
+  timestamp: string;
+  lastModified?: Date;
+};
+
+export type AgentSnapshotIterationJsonData = {
+  traceId: string;
+  iteration: number;
+  folderName: string;
+  timestamp: string;
+  jsonFiles: Record<string, any>;
+  totalFiles: number;
+};
